feat: replace use of enzyme with react-testing-library
- removed `cross-env` from test script
- moved CodeMirror mock to __mock__ file
