+++ packages/graphiql/src/components/__tests__/helpers/codeMirror.js
@@ -1,61 +1,3 @@
-export function MockCodeMirror(node, { value, ...options }) {
-  let _eventListeners = {};
-  const mockTextArea = document.createElement('textarea');
-  mockTextArea.className = 'mockCodeMirror';
-  mockTextArea.addEventListener('change', e => {
-    _emit('change', e);
-  });
-  mockTextArea.value = value;
-  node.appendChild(mockTextArea);
-
-  function _emit(event, data) {
-    if (_eventListeners[event]) {
-      _eventListeners[event](data);
-    }
-  }
-
-  return {
-    options: {
-      ...options,
-    },
-
-    on(event, handler) {
-      _eventListeners[event] = handler;
-    },
-
-    off(event) {
-      if (_eventListeners.hasOwnProperty(event)) {
-        const updatedEventListeners = {};
-        for (const e in _eventListeners) {
-          if (e !== event) {
-            updatedEventListeners[e] = _eventListeners[e];
-          }
-        }
-        _eventListeners = updatedEventListeners;
-      }
-    },
-
-    getValue() {
-      return mockTextArea.value;
-    },
-
-    setValue(newValue) {
-      mockTextArea.value = newValue;
-    },
-
-    setSize() {},
-
-    emit: _emit,
-  };
-}
-
-MockCodeMirror.defineExtension = () => {};
-MockCodeMirror.registerHelper = () => {};
-MockCodeMirror.defineOption = () => {};
-MockCodeMirror.signal = (mockCodeMirror, event, ...args) => {
-  mockCodeMirror.emit(event, ...args);
-};
-
 export const codeMirrorModules = [
   'codemirror/addon/hint/show-hint',
   'codemirror/addon/comment/comment',
